pyproject.toml
[project]
name = "campaign-command-center"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "openai>=1.30",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pydantic>=2.7",
    "numpy>=1.26",
    "scikit-learn>=1.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]
